+++ RestaurantReviews.Models/Review.cs
@@ -5,8 +5,8 @@ namespace RestaurantReviews.Models
     public class Review : IReview
     {
         public long Id { get; set; }
-        public int RestaurantId { get; set; }
-        public int UserId { get; set; }
+        public long RestaurantId { get; set; }
+        public long UserId { get; set; }
         public string Content { get; set; }
     }
 }
